server2/model/requirements.txt
Flask
Gunicorn
pandas
numpy
scikit-learn 
flask_cors
liac-arff